Add a method that will check computer guess
This method will return an array with values ranging from 0 to 2. These
values will indicate the computer's accuracy. 0 will be completely
incorrect. 1 will be a color in the wrong place. 2 will be a correct
color in the right place.

This way, the computer will be able to refine their guesses.